backend/requirements.txt
fastapi
uvicorn
numpy
Pillow
python-multipart
# tensorflow (optional, only needed if .h5 model is uploaded)
